Flask-app-1/requirements.txt
Flask==1.1.2
google-cloud-bigquery==2.13.1